requirements.txt
dataclasses_json
faiss-cpu
flytekit>=1.4.0
gitpython
langchain
openai
